+++ src/or_fail_functions.php
@@ -16,6 +16,7 @@ if (!function_exists('file_exists_or_fail')) {
      * Checks whether a file or directory exists and throws an exception if the
      *  file does not exist
      * @param string $filename Path to the file or directory
+     * @return void
      * @throws \ErrorException
      */
     function file_exists_or_fail($filename)
@@ -26,11 +27,30 @@ if (!function_exists('file_exists_or_fail')) {
     }
 }
 
+if (!function_exists('is_dir_or_fail')) {
+    /**
+     * Tells whether the filename is a directory and throws an exception if the
+     *  filename is not a directory
+     * @param string $filename Path to the directory
+     * @return void
+     * @throws \ErrorException
+     */
+    function is_dir_or_fail($filename)
+    {
+        file_exists_or_fail($filename);
+
+        if (!is_dir($filename)) {
+            throw new \ErrorException(sprintf('`%s` is not a directory', rtr($filename)));
+        }
+    }
+}
+
 if (!function_exists('is_readable_or_fail')) {
     /**
      * Tells whether a file exists and is readable and throws an exception if
      *  the file is not readable
      * @param string $filename Path to the file or directory
+     * @return void
      * @throws \ErrorException
      */
     function is_readable_or_fail($filename)
@@ -46,6 +66,7 @@ if (!function_exists('is_writable_or_fail')) {
      * Tells whether the filename is writable and throws an exception if the
      *  file is not writable
      * @param string $filename Path to the file or directory
+     * @return void
      * @throws \ErrorException
      */
     function is_writable_or_fail($filename)
@@ -55,4 +76,3 @@ if (!function_exists('is_writable_or_fail')) {
         }
     }
 }
-